RT-Ticket: 2586 RT-Status: resolved
As Alex suggested, it was only a matter of compress-plug assuming if you
were not read, you were write.  However, testing revealed that there was
a more fundamental problem with follow mode.  You could not cancel in
follow mode if there was no additional data coming in.  This is fixed
as well.