RT-Ticket: 3303 RT-Status: resolved
Well, it turned out to be a sort function problem after all.  The problem was that bk_vptr_cmp() returned the wrong sign if: the second vptr was shorter than the first AND the stored data was different.

This resolved the problem.  I am going to now double-check the BST bug that I 'found' to try and convince myself whether or not it really is needed.