<TRICKY>This is stupid code which will never be reached (which due to a gcc bug it does not know) to work around two different gcc optimization bugs:
warning: variable `foo' might be clobbered by `longjmp' or `vfork'

You are supposed to be able to make the variables volatile and
solve the problem, but it didn't work.  This *forces* the variables
into stack variables out of registers so they cannot be clobbered.
Sigh.</TRICKY>

This *could* be a sign that we have never actually used only nptl
before, since I *know* this machine is using it and has no backwards
compatibility.  This makes me sad about FC3's chances of using nptl.