GCC 4.0 support
* README.i18n: gcc-4 puts "const struct poptOption optionsTable[]"
in read-only code segment - hey - it's declared const, after all, and
this gets a SEGV when we localize one of the string pointers in it.
Don't declare it const, here, or in any of the umpteen dozen main()
functions that were cut&pasted from this

* include/libbk.h (BK_ENTRY_VOLATILE): BK_ENTRY_VOLATILE is
occasionally needed to prevent compiler warnings (errors) about
__bk_funinfo being clobbered by longjmp or vfork (or pthread_create).
<TRICKY>A special macro is needed because the __bk_funinfo pointer
variable *itself* (not the struct bk_funinfo it points to) must be
declared volatile - use cdecl to "explain" the crazy C syntax for this.

* lib/b_thread.c (bk_thread_create, bk_thread_kill_others): use it

* aclocal.m4 (AC_C_NO_POINTER_SIGN): check for ggc-4 -Wno-pointer-sign

* Make.config.in (NO_POINTER_SIGN_WARN): use autoconfig to set make var

* Make.variables (COMPILER_DEBUG+): use -Wno-pointer-sign if available

* lib/event-format/libevent.c (format_input_thread): BK_ENTRY_VOLATILE

* lib/jobstep/model.c (jobstep_input_event_thread): BK_ENTRY_VOLATILE