Bug 6898: Emergency mode off includes active blocking (runon + bk_timeout work)
Emergency mode off will now go to all sensor and request that all active
blocking actions be removed.  This uses the new runon program to execute
the commands on all sensors with some parallelism.  The new runon program
needed a better functioning timeout program.

The timeout program could exit with several return codes and had no way
of properly signalling to the callee whether the return code was due to
the child happening to exit with that code or due to a timeout.  Also,
the timeout name conflicted with a RHEL program of the same name (which
also did not have the needed semantics).  The timeout program was rename
to bk_timeout and all internal uses detected were renamed.  Timeout now
supports fractional second timeouts, a verbose mode which allows
disambiguation of return codes, and generally a far cleaner codepath.

Possible future extensions for bk_timeout might be to --wait for child to
die or to hard (-9) --killafter some specified number of seconds.

Runon should be used by more services which currently roll that kind of
thing by hand, poorly.  Possible extensions include syslog support
to log full status for posterity, and as documented in Bug 7161,
persistent commands for systems which are down