Bug 5190r: Time indexing moved from fs .aux to db
All of the .aux files which contain the information about the time range
stored in each pcap file have been obsoleted and moved to database tables,
specifically forens_raw and forens_vip.  In one specific test in a production
test system, the time required to search for a particular set of packets
went from 7 minutes to sub-second.

Note that this enhancement is strongly recommended to have a conversion
process executed on the forensics ring to convert the legacy .aux files
to database tables.  There are commands at the end of forensicsd.c
which will accomplish this task.

The ringdir code was refactored to pass information (specifically
the file number) around for use in the db code.

The awb-postgres code was refactored to expose the db helper functions
to non-worker users.