Typos
Break up bni_from_socket so that you can also create a bni from a sockaddr

Blocking version gethostbyfoo

Make mumurhash value fit within 32 bits when compiling on 32 bit machines

Set the primary address when adding the first address to a bni.

Fix a problem in both the DGRAM and STREAM connection functions where said
functions were taking improper advantage of the bug whose fix is mentioned
immediately above. Previously the primary address was (improperly) not yet
set on the first call to these functions, so now we have to make sure to
manually reset the address list on the first call.